requirements.txt
neuralFoil-torch
aerosandbox
matplotlib
